+++ main.go
@@ -40,7 +40,6 @@ import (
 	loggedcluster "github.com/giantswarm/logging-operator/pkg/logged-cluster"
 	loggingreconciler "github.com/giantswarm/logging-operator/pkg/logging-reconciler"
 	"github.com/giantswarm/logging-operator/pkg/reconciler"
-	alloysecret "github.com/giantswarm/logging-operator/pkg/resource/alloy-secret"
 	grafanaagentconfig "github.com/giantswarm/logging-operator/pkg/resource/grafana-agent-config"
 	grafanaagentsecret "github.com/giantswarm/logging-operator/pkg/resource/grafana-agent-secret"
 	grafanadatasource "github.com/giantswarm/logging-operator/pkg/resource/grafana-datasource"
@@ -162,10 +161,6 @@ func main() {
 		Client: mgr.GetClient(),
 	}
 
-	alloySecret := alloysecret.Reconciler{
-		Client: mgr.GetClient(),
-	}
-
 	podLogs := podlogs.Reconciler{
 		Client: mgr.GetClient(),
 	}
@@ -189,7 +184,6 @@ func main() {
 			&loggingConfig,
 			&grafanaAgentSecret,
 			&grafanaAgentConfig,
-			&alloySecret,
 			&podLogs,
 		},
 	}